fix: refactor to proxy all Notion endpoints through lambdas
