+++ next.config.js
@@ -1,8 +1,9 @@
-{
-  module.exports = {
-    target: "serverless",
-    experimental: {
-      ampBindInitData: true
-    }
-  };
-}
+module.exports = {
+  target: "serverless",
+  experimental: {
+    ampBindInitData: true
+  },
+  env: {
+    DEV: process.env.NODE_ENV !== 'production',
+  }
+};
