+++ notion/utils/index.js
@@ -1,0 +1,21 @@
+async function getError(res) {
+  return `Notion API error (${res.status}) \n${getJSONHeaders(
+    res
+  )}\n ${await getBodyOrNull(res)}`;
+}
+
+function getJSONHeaders(res) {
+  return JSON.stringify(res.headers.raw());
+}
+
+function getBodyOrNull(res) {
+  try {
+    return res.text();
+  } catch (err) {
+    return null;
+  }
+}
+
+module.exports = {
+  getError,
+}
