+++ package.json
@@ -1,6 +1,8 @@
 {
   "dependencies": {
+    "body-parser": "^1.19.0",
     "color": "^3.1.1",
+    "express": "^4.16.4",
     "fetch-retry": "^2.0.0",
     "isomorphic-unfetch": "^3.0.0",
     "next": "^8.1.0",
